requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.7.0
django-unfold==0.55.2
pillow==10.4.0
sqlparse==0.5.0
typing_extensions==4.12.2
